Some small fixes after porting to Java
